backend/webapp/base_requirements.txt
Django==1.8
djangorestframework-3.5.3